Former-commit-id: 12c747d0bd708405d44a5ca84292e508f3c5fa0a

+++ src/scene_server/topo_server/core/operation/inst_set.go
@@ -21,6 +21,7 @@ import (
 	"configcenter/src/common/condition"
 	"configcenter/src/common/mapstr"
 	"configcenter/src/common/metadata"
+	"configcenter/src/common/util"
 	"configcenter/src/scene_server/topo_server/core/inst"
 	"configcenter/src/scene_server/topo_server/core/model"
 	"configcenter/src/scene_server/topo_server/core/types"
@@ -80,11 +81,77 @@ func (s *set) CreateSet(params types.ContextParams, obj model.Object, bizID int6
 	data.Set(common.BKAppIDField, bizID)
 
 	if !data.Exists(common.BKDefaultField) {
-		data.Set(common.BKDefaultField, 0)
+		data.Set(common.BKDefaultField, common.DefaultFlagDefaultValue)
 	}
 
-	// data.Set(common.CreateTimeField, util.GetCurrentTimeStr())
-	return s.inst.CreateInst(params, obj, data)
+	setTemplate := metadata.SetTemplate{}
+	// validate foreign key
+	if setTemplateIDIf, ok := data[common.BKSetTemplateIDField]; ok == true {
+		setTemplateID, err := util.GetInt64ByInterface(setTemplateIDIf)
+		if err != nil {
+			blog.Errorf("parse set_template_id field into int failed, id: %+v, rid: %s", setTemplateIDIf, params.ReqID)
+			err := params.Err.CCErrorf(common.CCErrCommParamsInvalid, params.Lang.Language("set_property_set_template_id"))
+			return nil, err
+		}
+		if setTemplateID != common.SetTemplateIDNotSet {
+			st, err := s.clientSet.CoreService().SetTemplate().GetSetTemplate(params.Context, params.Header, bizID, setTemplateID)
+			if err != nil {
+				err := params.Err.CCErrorf(common.CCErrCommParamsInvalid, params.Lang.Language("set_property_set_template_id"))
+				return nil, err
+			}
+			setTemplate = st
+		}
+	} else {
+		data[common.BKSetTemplateIDField] = common.SetTemplateIDNotSet
+	}
+
+	// TODO: run in transaction
+	data.Set(common.BKSetTemplateIDField, setTemplate.ID)
+	data.Remove(common.MetadataField)
+	setInstance, err := s.inst.CreateInst(params, obj, data)
+	if err != nil {
+		blog.Errorf("create set instance failed, object: %+v, data: %+v, err: %s, rid: %s", obj, data, err.Error(), params.ReqID)
+		return setInstance, err
+	}
+	if setTemplate.ID == 0 {
+		return setInstance, nil
+	}
+
+	setID, err := setInstance.GetInstID()
+	if err != nil {
+		blog.Errorf("create set instance success, but read instance id field failed, bizID: %d, setInstance: %+v, err: %s, rid: %s", bizID, setInstance, err.Error(), params.ReqID)
+		return setInstance, err
+	}
+
+	// set create by template should create module at the same time
+	serviceTemplates, err := s.clientSet.CoreService().SetTemplate().ListSetTplRelatedSvcTpl(params.Context, params.Header, bizID, setTemplate.ID)
+	if err != nil {
+		blog.Errorf("create set failed, ListSetTplRelatedSvcTpl failed, bizID: %d, setTemplateID: %d, err: %s, rid: %s", bizID, setTemplate.ID, err.Error(), params.ReqID)
+		return setInstance, err
+	}
+
+	moduleObj, err := s.obj.FindSingleObject(params, common.BKInnerObjIDModule)
+	if nil != err {
+		blog.Errorf("[operation-set] failed to find module object, error info is %s, rid: %s", err.Error(), params.ReqID)
+		return setInstance, err
+	}
+	for _, serviceTemplate := range serviceTemplates {
+		createModuleParam := map[string]interface{}{
+			common.BKModuleNameField:        serviceTemplate.Name,
+			common.BKServiceTemplateIDField: serviceTemplate.ID,
+			common.BKSetTemplateIDField:     setTemplate.ID,
+			common.BKParentIDField:          setID,
+			common.BKServiceCategoryIDField: serviceTemplate.ServiceCategoryID,
+			common.BKAppIDField:             bizID,
+		}
+		_, err := s.module.CreateModule(params, moduleObj, bizID, setID, createModuleParam)
+		if err != nil {
+			blog.Errorf("create set instance failed, create module instance failed, bizID: %d, setID: %d, param: %+v, err: %s, rid: %s", bizID, setID, createModuleParam, err.Error(), params.ReqID)
+			return setInstance, err
+		}
+	}
+
+	return setInstance, nil
 }
 
 func (s *set) DeleteSet(params types.ContextParams, setModel model.Object, bizID int64, setIDS []int64) error {
@@ -121,6 +188,12 @@ func (s *set) DeleteSet(params types.ContextParams, setModel model.Object, bizID
 		return params.Err.New(common.CCErrTopoSetDeleteFailed, err.Error())
 	}
 
+	// clear set template sync status
+	if ccErr := s.clientSet.CoreService().SetTemplate().DeleteSetTemplateSyncStatus(params.Context, params.Header, bizID, setIDS); ccErr != nil {
+		blog.Errorf("[operation-set] failed to delete set template sync status failed, bizID: %d, setIDs: %+v, err: %s, rid: %s", bizID, setIDS, ccErr.Error(), params.ReqID)
+		return ccErr
+	}
+
 	// clear the sets
 	return s.inst.DeleteInst(params, setModel, setCond, false)
 }
@@ -135,6 +208,10 @@ func (s *set) UpdateSet(params types.ContextParams, data mapstr.MapStr, obj mode
 
 	innerCond.Field(common.BKAppIDField).Eq(bizID)
 	innerCond.Field(common.BKSetIDField).Eq(setID)
+
+	data.Remove(common.MetadataField)
+	data.Remove(common.BKAppIDField)
+	data.Remove(common.BKSetIDField)
 
 	return s.inst.UpdateInst(params, data, obj, innerCond, setID)
 }
